daemon, polkit: update based on John Lenton's review
* rename polkit.CheckAuthorization to CheckAuthorizationForPid, and make
  it return (bool, error).  If we need to check other subject types in
  future, we can add extra entry points.

* If the response says the user dismissed the auth dialog, return
  ErrDismissed.  In daemon, don't log failures due to this error.

* If the response says a challenge is required, rewturn
  ErrInteractionRequired.

* Use uint32 to represent process IDs: this is what is used at the D-Bus
  level, and even 64-bit Linux limits IDs to 2^22.  This leads to fewer
  casts.

+++ daemon/api_test.go
@@ -4547,7 +4547,7 @@ func (s *postCreateUserSuite) SetUpTest(c *check.C) {
 	s.apiBaseSuite.SetUpTest(c)
 
 	s.daemon(c)
-	postCreateUserUcrednetGet = func(string) (uint64, uint32, error) {
+	postCreateUserUcrednetGet = func(string) (uint32, uint32, error) {
 		return 100, 0, nil
 	}
 	s.mockUserHome = c.MkDir()
@@ -4907,7 +4907,7 @@ func (s *postCreateUserSuite) TestPostCreateUserFromAssertionAllKnownClassicErro
 
 	s.makeSystemUsers(c, []map[string]interface{}{goodUser})
 
-	postCreateUserUcrednetGet = func(string) (uint64, uint32, error) {
+	postCreateUserUcrednetGet = func(string) (uint32, uint32, error) {
 		return 100, 0, nil
 	}
 	defer func() {
